+++ LawProject/LawProject.Domain/Entities/PostTag.cs
@@ -6,7 +6,7 @@ namespace LawProject.Domain.Entities
 {
     public class PostTag
     {
-        public int PostID { set; get; }
+        public long PostID { set; get; }
         public string TagID { set; get; }
         public virtual Post Post { set; get; }
         public virtual Tag Tag { set; get; }
